refactor: run ng update for core
